chore: Use the split components
