--- webapp/src/components/AssetPage/AssetHistory/HistoryTable/HistoryTable.module.css
@@ -1,24 +1,0 @@
-.main {
-  position: relative;
-}
-
-.main tbody.isLoading tr {
-  opacity: 0.5;
-}
-
-.mobileHistoryTable {
-  margin-top: 16px;
-  margin-bottom: 96px;
-  font-size: 15px;
-}
-
-.mobileHistoryTableRow {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 8px;
-}
-
-.mobileHistoryRowStartedAt {
-  color: var(--secondary-text);
-}
