--- webapp/src/components/AssetPage/AssetHistory/HistoryTable/index.ts
@@ -1,2 +1,0 @@
-import HistoryTable from './HistoryTable'
-export { HistoryTable }
