--- webapp/src/components/AssetPage/AssetHistory/RentalHistory/RentalHistory.types.ts
@@ -1,8 +1,0 @@
-import { NFT } from '../../../../modules/nft/types'
-
-export type Props = {
-  asset: NFT
-}
-
-export type MapStateProps = {}
-export type MapDispatchProps = {}
