--- webapp/src/components/AssetPage/AssetHistory/RentalHistory/index.ts
@@ -1,2 +1,0 @@
-import RentalHistory from './RentalHistory'
-export { RentalHistory }
