--- webapp/src/components/AssetPage/AssetHistory/TransactionHistory/TransactionHistory.types.ts
@@ -1,8 +1,0 @@
-import { Asset } from '../../../../modules/asset/types'
-
-export type Props = {
-  asset: Asset
-}
-
-export type MapStateProps = {}
-export type MapDispatchProps = {}
