--- webapp/src/components/AssetPage/AssetHistory/TransactionHistory/index.ts
@@ -1,2 +1,0 @@
-import TransactionHistory from './TransactionHistory'
-export { TransactionHistory }
