--- webapp/src/components/AssetPage/AssetHistory/index.ts
@@ -1,2 +1,0 @@
-export * from './RentalHistory'
-export * from './TransactionHistory'
