+++ webapp/src/components/AssetPage/EstateDetail/EstateDetail.tsx
@@ -14,7 +14,7 @@ import { BidList } from '../BidList'
 import { JumpIn } from '../JumpIn'
 import { ProximityHighlights } from '../ProximityHighlights'
 import BaseDetail from '../BaseDetail'
-import { TransactionHistory } from '../AssetHistory'
+import { TransactionHistory } from '../TransactionHistory'
 import { SaleRentActionBox } from '../SaleRentActionBox'
 import { Props } from './EstateDetail.types'
 import { ParcelCoordinates } from './ParcelCoordinates'
