+++ webapp/src/components/AssetPage/ParcelDetail/ParcelDetail.tsx
@@ -13,7 +13,8 @@ import Expiration from '../Expiration'
 import { Actions } from '../Actions'
 import { BidList } from '../BidList'
 import { JumpIn } from '../JumpIn'
-import { TransactionHistory, RentalHistory } from '../AssetHistory'
+import { TransactionHistory } from '../TransactionHistory'
+import { RentalHistory } from '../RentalHistory'
 import { ProximityHighlights } from '../ProximityHighlights'
 import BaseDetail from '../BaseDetail'
 import { SaleRentActionBox } from '../SaleRentActionBox'
